Document ZoomSlider, ZoomControl, LayerList and LayerListItem
